% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatostone.h.R
\name{dataTOSTone}
\alias{dataTOSTone}
\title{TOST One Sample T-Test}
\usage{
dataTOSTone(data, vars, mu = 0, lowEqBD = -0.5, highEqBD = 0.5,
  alpha = 0.05, desc = FALSE, plots = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{a vector of strings naming variables of interest in \code{data}}

\item{mu}{a number (default: 0) to compare against}

\item{lowEqBD}{a number (default: -0.5), the lower equivalence bound in 
Cohen's D}

\item{highEqBD}{a number (default: 0.5), the upper equivalence bound in 
Cohen's D}

\item{alpha}{a number between 0 and 1 (default: 0.05) specifying the alpha 
level}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive 
statistics}

\item{plots}{\code{TRUE} or \code{FALSE} (default), provide plots}
}
\description{
TOST One Sample T-Test
}

