% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatostr.h.R
\name{dataTOSTr}
\alias{dataTOSTr}
\title{TOST Correlation}
\usage{
dataTOSTr(data, pairs, lowEqBR = -0.3, highEqBR = 0.3, alpha = 0.05,
  desc = FALSE, plots = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{pairs}{a list of vectors of strings naming variables to pair from 
\code{data}}

\item{lowEqBR}{a number (default: -0.3), the lower equivalence bound in r}

\item{highEqBR}{a number (default: 0.3), the upper equivalence bound in r}

\item{alpha}{a number between 0 and 1 (default: 0.05) specifying the alpha 
level}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive 
statistics}

\item{plots}{\code{TRUE} or \code{FALSE} (default), provide plots}
}
\description{
TOST Correlation
}
