% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatosttwo.h.R
\name{dataTOSTtwo}
\alias{dataTOSTtwo}
\title{TOST Independent Samples T-Test}
\usage{
dataTOSTtwo(data, deps, group, eqVar = FALSE, lowEqBD = -0.5,
  highEqBD = 0.5, alpha = 0.05, desc = FALSE, plots = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{deps}{a vector of strings naming dependent variables in \code{data}}

\item{group}{a string naming the grouping variable in \code{data}; must 
have two levels}

\item{eqVar}{\code{TRUE} or \code{FALSE} (default), assume equal variances}

\item{lowEqBD}{a number (default: -0.5), the lower equivalence bound in 
Cohen's D}

\item{highEqBD}{a number (default: 0.5), the upper equivalence bound in 
Cohen's D}

\item{alpha}{a number between 0 and 1 (default: 0.05) specifying the alpha 
level}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive 
statistics}

\item{plots}{\code{TRUE} or \code{FALSE} (default), provide plots}
}
\description{
TOST Independent Samples T-Test
}
