% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_calcs.R
\name{.fisher_z_se}
\alias{.fisher_z_se}
\title{Analytical SE on the Fisher z scale for a given correlation method}
\usage{
.fisher_z_se(r_star, n, method)
}
\arguments{
\item{r_star}{Correlation estimate (can be a vector for bootstrap replicates)}

\item{n}{Sample size}

\item{method}{One of "pearson", "kendall", "spearman"}
}
\value{
SE on the Fisher z scale
}
\description{
Analytical SE on the Fisher z scale for a given correlation method
}
\keyword{internal}
