% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_r_paired.R
\name{extract_r_paired}
\alias{extract_r_paired}
\title{Extract Paired Correlation}
\usage{
extract_r_paired(m1, sd1, m2, sd2 = NULL, n, tstat = NULL, pvalue = NULL)
}
\arguments{
\item{m1}{mean of group 1}

\item{sd1}{standard deviation of group 1}

\item{m2}{mean of group 2}

\item{sd2}{standard deviation of group 2}

\item{n}{Sample size (number of pairs)}

\item{tstat}{The t-value from a paired samples t-test}

\item{pvalue}{The two-tailed p-value from a paired samples t-test}
}
\value{
An estimate of the correlation.
}
\description{
A function for estimating the correlation from a paired samples t-test.
Useful for tsum_TOST when the correlation is not available.
}
\references{
Lajeunesse, M. J. (2011). On the meta‐analysis of response ratios for studies with correlated and multi‐group designs. Ecology, 92(11), 2049-2055
}
