% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_helpers.R
\name{htest-helpers}
\alias{htest-helpers}
\alias{df_htest}
\alias{explain_htest}
\title{Helpers for \code{htest} objects}
\usage{
df_htest(htest, test_statistics = TRUE, show_ci = TRUE, extract_names = TRUE)

explain_htest(htest)
}
\arguments{
\item{htest}{A S3 object of the class \code{htest}}

\item{test_statistics}{A logical variable to display the test statistics.}

\item{show_ci}{A logical variable to display the confidence interval.}

\item{extract_names}{A logical variable to take the names from the S3 object (i.e., statistic for \code{t.test} would be "t")}
}
\description{
Functions to help interpret or display objects of the class \code{htest}
}
\examples{
# To be added

}
