% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_TOST.R
\name{log_TOST}
\alias{log_TOST}
\alias{log_TOST.default}
\alias{log_TOST.formula}
\title{TOST with log transformed t-tests}
\usage{
log_TOST(
  x,
  ...,
  hypothesis = "EQU",
  paired = FALSE,
  var.equal = FALSE,
  eqb = 1.25,
  alpha = 0.05,
  null = 1
)

\method{log_TOST}{default}(
  x,
  y = NULL,
  hypothesis = c("EQU", "MET"),
  var.equal = FALSE,
  paired = FALSE,
  eqb = 1.25,
  alpha = 0.05,
  null = 1,
  ...
)

\method{log_TOST}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods.}

\item{hypothesis}{'EQU' for equivalence (default), or 'MET' for minimal effects test, the alternative hypothesis.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{eqb}{Equivalence bound; default is 1.25 (FDA guidelines). Can provide 1 value (reciprical value is taken as the lower bound) or 2 specific values that represent the upper and lower equivalence bounds.}

\item{alpha}{alpha level (default = 0.05)}

\item{null}{Null hypothesis value for two-tailed test (default is 1).}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}
}
\value{
An S3 object of class
  \code{"TOSTt"} is returned containing the following slots:
\describe{
  \item{\code{"TOST"}}{A table of class \code{"data.frame"} containing two-tailed t-test and both one-tailed results.}
  \item{\code{"eqb"}}{A table of class \code{"data.frame"} containing equivalence bound settings.}
  \item{\code{"effsize"}}{ table of class \code{"data.frame"} containing effect size estimates}
  \item{\code{"hypothesis"}}{String stating the hypothesis being tested}
  \item{\code{"smd"}}{List containing the results of the standardized mean difference calculations (e.g., Cohen's d).Items include: d (estimate), dlow (lower CI bound), dhigh (upper CI bound), d_df (degrees of freedom for SMD), d_sigma (SE), d_lambda (non-centrality), J (bias correction), smd_label (type of SMD), d_denom (denominator calculation)}
  \item{\code{"alpha"}}{Alpha level set for the analysis.}
  \item{\code{"method"}}{Type of t-test.}
  \item{\code{"decision"}}{List included text regarding the decisions for statistical inference.}
}
}
\description{
A function for TOST with all types of t-tests.
}
