% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOST.R
\name{powerTOST}
\alias{powerTOST}
\title{Function to perform power analysis for TOST}
\usage{
powerTOST(alpha, statistical_power, low_eqbound_d, high_eqbound_d, epsilon)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound_d}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's d)}

\item{high_eqbound_d}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's d)}

\item{epsilon}{By default, true effect is assumed to be 0. If you want to perform an equivalence test when expecting a non-zero effect, specify the expected standardized effect size in Cohen's d as epsilon (e.g., 0.05)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of observations needed in each group
}
\description{
Function to perform power analysis for TOST
}
\examples{
powerTOST(alpha=0.05, beta=0.2, low_eqbound_d=-0.4, high_eqbound_d=0.4, epsilon=0.1) #Returns n (for each condition)
}

