% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOSTpaired.R
\name{powerTOSTpaired}
\alias{powerTOSTpaired}
\title{Power analysis for TOST for dependent t-test (Cohen's dz)}
\usage{
powerTOSTpaired(alpha, statistical_power, low_eqbound_dz, high_eqbound_dz)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound_dz}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's dz)}

\item{high_eqbound_dz}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's dz)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of pairs needed
}
\description{
Power analysis for TOST for dependent t-test (Cohen's dz)
}
\section{References}{

Chow, S.-C., Wang, H., & Shao, J. (2007). Sample Size Calculations in Clinical Research, Second Edition - CRC Press Book. Formula 3.1.9
}
\examples{
powerTOSTpaired(alpha=0.05,statistical_power=0.8,low_eqbound_dz=-0.3, high_eqbound_dz=0.3)
}

