% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOSTr.R
\name{powerTOSTr}
\alias{powerTOSTr}
\title{Function to perform power analysis for TOST}
\usage{
powerTOSTr(alpha, statistical_power, low_eqbound_r, high_eqbound_r, epsilon)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound_r}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's dz)}

\item{high_eqbound_r}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's dz)}

\item{epsilon}{By default, true effect is assumed to be 0. If you want to perform an equivalence test when expecting a non-zero effect, specify the expected standardized effect size in Cohen's dz as epsilon (e.g., 0.05)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of pairs needed
}
\description{
Function to perform power analysis for TOST
}
\examples{
powerTOSTr(alpha=0.05, statistical_power=0.8, low_eqbound_r=-0.1, high_eqbound_r=0.1)
}

