% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOSTraw.R
\name{powerTOSTraw}
\alias{powerTOSTraw}
\title{Function to perform power analysis for TOST}
\usage{
powerTOSTraw(alpha, statistical_power, low_eqbound, high_eqbound, sdpooled,
  epsilon)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints)}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in raw scale units (e.g., scalepoints)}

\item{epsilon}{By default, true effect is assumed to be 0. If you want to perform an equivalence test when expecting a non-zero effect, specify the expected standardized effect size raw scale units (e.g., scalepoints) as epsilon (e.g., 0.50)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of observations needed in each group
}
\description{
Function to perform power analysis for TOST
}
\examples{
powerTOSTraw(alpha=0.05, statistical_power=0.8, low_eqbound=-200, high_eqbound=200, sdpooled=350, epsilon=10)
}

