% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTdz.R
\name{TOSTdz}
\alias{TOSTdz}
\title{TOST function for a dependent t-test}
\usage{
TOSTdz(n, m1, m2, sd1, sd2, r12, low_eqbound_dz, high_eqbound_dz, alpha)
}
\arguments{
\item{n}{sample size (pairs)}

\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{r12}{correlation of dependent variable between group 1 and  group 2}

\item{low_eqbound_dz}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's dz)}

\item{high_eqbound_dz}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's dz)}

\item{alpha}{alpha level (default = 0.05)}
}
\value{
Returns dataframe with t-value, degrees of freedom, p-value, and lower and upper limit of confidence interval
}
\description{
TOST function for a dependent t-test
}
\examples{
TOSTdz(n=65,m1=5.830769,m2=5.7538462,sd1=1.1668498,sd2=1.2994081,r12=0.744992,low_eqbound_dz=-0.337341706,high_eqbound_dz=0.337341706, alpha=0.05)
}

