% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTER-package.R
\docType{package}
\name{TOSTER-package}
\alias{TOSTER}
\alias{TOSTER-package}
\title{TOSTER: Two One-Sided Tests (TOST) Equivalence Testing}
\description{
Two one-sided tests (TOST) procedure to test equivalence for t-tests, correlations, differences between proportions, and meta-analyses, including power analysis for t-tests and correlations. Allows you to specify equivalence bounds in raw scale units or in terms of effect sizes. See: Lakens (2017) \doi{10.1177/1948550617697177}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aaroncaldwell.us/TOSTERpkg/}
}

}
\author{
\strong{Maintainer}: Aaron Caldwell \email{arcaldwell49@gmail.com}

Authors:
\itemize{
  \item Daniel Lakens \email{D.Lakens@tue.nl}
}

}
\keyword{internal}
