% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTmeta.R
\name{TOSTmeta}
\alias{TOSTmeta}
\title{TOST function for meta-analysis}
\usage{
TOSTmeta(ES, var, se, low_eqbound_d, high_eqbound_d, alpha)
}
\arguments{
\item{ES}{meta-analytic effect size}

\item{var}{meta-analytic variance}

\item{se}{standard error}

\item{low_eqbound_d}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's d)}

\item{high_eqbound_d}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's d)}

\item{alpha}{alpha level (default = 0.05)}
}
\value{
Returns dataframe with Z-value, p-value, and lower and upper limit of confidence interval
}
\description{
TOST function for meta-analysis
}
\examples{
TOSTmeta(ES=0.12, var=0.0081, se=0.09, low_eqbound_d=-0.2, high_eqbound_d=0.2, alpha=0.05)
}

