% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTone.R
\name{TOSTone}
\alias{TOSTone}
\title{TOSTone function for a one-sample t-test}
\usage{
TOSTone(m, mu, sd, n, low_eqbound_d, high_eqbound_d, alpha)
}
\arguments{
\item{m}{mean}

\item{mu}{value to compare against}

\item{sd}{standard deviation}

\item{n}{sample size}

\item{low_eqbound_d}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's d)}

\item{high_eqbound_d}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's d)}

\item{alpha}{alpha level (default = 0.05)}
}
\value{
Returns dataframe with t-value, degrees of freedom, p-value, and lower and upper limit of confidence interval
}
\description{
TOSTone function for a one-sample t-test
}
\examples{
TOST(m=0.54,mu=0.5,sd=1.2,n=100,low_eqbound_d=-0.3, high_eqbound_d=0.3, alpha=0.05)
}

