% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTr.R
\name{TOSTr}
\alias{TOSTr}
\title{TOST function for a correlations}
\usage{
TOSTr(n, r, low_eqbound_r, high_eqbound_r, alpha)
}
\arguments{
\item{n}{number of pairs of observations}

\item{r}{observed correlation}

\item{low_eqbound_r}{lower equivalence bounds (e.g., -0.3) expressed in a correlation effect size}

\item{high_eqbound_r}{upper equivalence bounds (e.g., 0.3) expressed in a correlation effect size}

\item{alpha}{alpha level (default = 0.05)}
}
\value{
Returns dataframe with p-value, and lower and upper limit of confidence interval
}
\description{
TOST function for a correlations
}
\examples{
TOSTr(n=100, r = 0.02, low_eqbound_r=-0.3, high_eqbound_r=0.3, alpha=0.05)
}

