% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTraw.R
\name{TOSTraw}
\alias{TOSTraw}
\title{TOST raw function}
\usage{
TOSTraw(m1, m2, sd1, sd2, n1, n2, low_eqbound, high_eqbound, alpha, var.equal)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{n1}{sample size in group 1}

\item{n2}{sample size in group 2}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints)}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in raw scale units (e.g., scalepoints)}

\item{alpha}{alpha level (default = 0.05)}

\item{var.equal}{logical variable indicating whether equal variances assumption is assumed to be TRUE or FALSE. Defaults to FALSE.}
}
\value{
Returns dataframe with t-value, degrees of freedom, p-value, and lower and upper limit of confidence interval
}
\description{
TOST raw function
}
\examples{
TOSTraw(m1=7.83,m2=7.98,sd1=1.21,sd2=1.29,n1=400,n2=400,low_eqbound=-0.25, high_eqbound=0.25,alpha=0.05)
}

