% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTtwo.raw.R
\name{TOSTtwo.raw}
\alias{TOSTtwo.raw}
\title{TOST function for an independent t-test (raw scores)}
\usage{
TOSTtwo.raw(m1, m2, sd1, sd2, n1, n2, low_eqbound, high_eqbound, alpha,
  var.equal)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{n1}{sample size in group 1}

\item{n2}{sample size in group 2}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in raw scale units (e.g., scalepoints)}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in raw scale units (e.g., scalepoints)}

\item{alpha}{alpha level (default = 0.05)}

\item{var.equal}{logical variable indicating whether equal variances assumption is assumed to be TRUE or FALSE. Defaults to FALSE.}
}
\value{
Returns TOST t-value 1, TOST p-value 1, TOST t-value 2, TOST p-value 2, degrees of freedom, low equivalence bound, high equivalence bound, Lower limit confidence interval TOST, Upper limit confidence interval TOST
}
\description{
TOST function for an independent t-test (raw scores)
}
\section{References}{

Berger, R. L., & Hsu, J. C. (1996). Bioequivalence Trials, Intersection-Union Tests and Equivalence Confidence Sets. Statistical Science, 11(4), 283-302. 
Gruman, J. A., Cribbie, R. A., & Arpin-Cribbie, C. A. (2007). The effects of heteroscedasticity on tests of equivalence. Journal of Modern Applied Statistical Methods, 6(1), 133-140, formula for Welch's t-test on page 135
}
\examples{
TOSTtwo.raw(m1=7.83,m2=7.98,sd1=1.21,sd2=1.29,n1=400,n2=400,low_eqbound=-0.25, high_eqbound=0.25,alpha=0.05)
}

