% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_proportions.R
\name{twoprop_test}
\alias{twoprop_test}
\title{Testing for differences in proportions in 2 groups}
\usage{
twoprop_test(
  p1,
  p2,
  n1,
  n2,
  null = 0,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater", "equivalence", "minimal.effect")
)
}
\arguments{
\item{p1, p2}{Proportions in each respective group.}

\item{n1, n2}{sample size in each respective group.}

\item{null}{a number indicating the null hypothesis of the difference in proportions between two groups.}

\item{alpha}{alpha level (default = 0.05)}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", "less", "equivalence" (TOST), or "minimal.effect" (TOST). You can specify just the initial letter.}
}
\value{
An S3 object of the class \code{htest}.
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
This is a hypothesis testing function that mimics \link{prop.test}, but focuses only on testing differences in proportions between two groups.
This function utilizes a z-test with unpooled variance estimates.
}
\references{
Tunes da Silva, G., Logan, B. R., & Klein, J. P. (2008). Methods for Equivalence and Noninferiority Testing. Biology of Blood Marrow Transplant, 15(1 Suppl), 120-127.

Yin, G. (2012). Clinical Trial Design: Bayesian and Frequentist Adaptive Methods. Hoboken, New Jersey: John Wiley & Sons, Inc.
}
