% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_TOST.R
\name{wilcox_TOST}
\alias{wilcox_TOST}
\alias{wilcox_TOST.default}
\alias{wilcox_TOST.formula}
\title{TOST with Wilcoxon-Mann-Whitney tests}
\usage{
wilcox_TOST(
  x,
  ...,
  hypothesis = "EQU",
  paired = FALSE,
  eqb,
  low_eqbound,
  high_eqbound,
  ses = "rb",
  alpha = 0.05
)

\method{wilcox_TOST}{default}(
  x,
  y = NULL,
  hypothesis = "EQU",
  paired = FALSE,
  eqb,
  low_eqbound,
  high_eqbound,
  ses = c("rb", "odds", "logodds", "cstat"),
  alpha = 0.05,
  mu = 0,
  ...
)

\method{wilcox_TOST}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods.}

\item{hypothesis}{'EQU' for equivalence (default), or 'MET' for minimal effects test.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{eqb}{Equivalence bound. Can provide 1 value (symmetric bound) or 2 specific values that represent the lower and upper equivalence bounds.
Like the \code{mu} argument, \code{eqb} is specified on the \strong{raw scale} of the original data (e.g., the scale of the median or location shift).
This parameter is independent of the \code{ses} argument, which only affects the type of standardized effect size that is reported.}

\item{low_eqbound}{lower equivalence bounds (deprecated, use \code{eqb} instead).}

\item{high_eqbound}{upper equivalence bounds (deprecated, use \code{eqb} instead).}

\item{ses}{Standardized effect size. Default is "rb" for rank-biserial
correlation. Options also include "cstat" for concordance probability, or
"odds" for Wilcoxon-Mann-Whitney odds (otherwise known as Agresti's
generalized odds ratio). Note that \code{ses} only determines which effect size is calculated and does not affect the equivalence bounds (\code{eqb}).}

\item{alpha}{alpha level (default = 0.05)}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{mu}{number indicating the value around which (a-)symmetry (for
one-sample or paired samples) or shift (for independent samples) is to be
estimated. See \link[stats:wilcox.test]{stats::wilcox.test}.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}
}
\value{
An S3 object of class
\code{"TOSTnp"} is returned containing the following slots:
\itemize{
\item "TOST": A table of class \code{"data.frame"} containing two-tailed wilcoxon signed rank test and both one-tailed results.
\item "eqb": A table of class \code{"data.frame"} containing equivalence bound settings.
\item "effsize":  table of class \code{"data.frame"} containing effect size estimates.
\item "hypothesis": String stating the hypothesis being tested.
\item "smd": List containing information on standardized effect size.
\item "alpha": Alpha level set for the analysis.
\item "method": Type of non-parametric test.
\item "decision": List included text regarding the decisions for statistical inference.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A function for TOST using the non-parametric methods of the Wilcoxon-Mann-Whitney family of tests.
This function uses the normal approximation and applies a continuity correction automatically.
}
\details{
For details on the calculations in this function see \code{vignette("robustTOST")}.

If only x is given, or if both x and y are given and paired is TRUE,
a Wilcoxon signed rank test of the null that the distribution of x (in the one sample case)
or of x - y (in the paired two sample case) is symmetric about mu is performed.

Otherwise, if both x and y are given and paired is FALSE,
a Wilcoxon rank sum test (equivalent to the Mann-Whitney test: see the Note) is carried out.
In this case, the null hypothesis is that the distributions of x and y differ by a location shift.
}
\section{References}{

David F. Bauer (1972). Constructing confidence sets using rank statistics. Journal of the American Statistical Association 67, 687–690. doi: 10.1080/01621459.1972.10481279.

Myles Hollander and Douglas A. Wolfe (1973). Nonparametric Statistical Methods. New York: John Wiley & Sons. Pages 27–33 (one-sample), 68–75 (two-sample). Or second edition (1999).
}

\examples{
data(mtcars)
wilcox_TOST(mpg ~ am,
data = mtcars,
eqb = 3)
}
\seealso{
Other Robust tests: 
\code{\link{boot_log_TOST}()},
\code{\link{boot_t_TOST}()},
\code{\link{boot_t_test}()},
\code{\link{brunner_munzel}()},
\code{\link{log_TOST}()}

Other TOST: 
\code{\link{boot_log_TOST}()},
\code{\link{boot_t_TOST}()},
\code{\link{simple_htest}()},
\code{\link{t_TOST}()},
\code{\link{tsum_TOST}()}
}
\concept{Robust tests}
\concept{TOST}
