% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{z_cor_test}
\alias{z_cor_test}
\title{Test for Association/Correlation Between Paired Samples}
\usage{
z_cor_test(
  x,
  y,
  alternative = c("two.sided", "less", "greater"),
  method = c("pearson", "kendall", "spearman"),
  alpha = 0.05,
  null = 0,
  TOST = FALSE
)
}
\arguments{
\item{x, y}{numeric vectors of data values. x and y must have the same length.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{null}{a number indicating the null hypothesis. Default is a correlation of zero.}

\item{TOST}{logical indicator (default = FALSE) to perform two one-sided tests of equivalence (TOST). Minimal effects testing not currently available. If specified, alternative is ignored.
 @return A list with class "htest" containing the following components:
\describe{
  \item{\code{"statistic"}}{z-score}
  \item{\code{"p.value"}}{the p-value of the test.}
  \item{\code{"estimate"}}{the estimated measure of association, with name "cor", "tau", or "rho" corresponding to the method employed.}
  \item{\code{"null.value"}}{the value of the association measure under the null hypothesis.}
  \item{\code{"alternative"}}{character string indicating the alternative hypothesis (the value of the input argument alternative). Possible values are "greater", "less", or "two-sided".}
  \item{\code{"method"}}{a character string indicating how the association was measured.}
  \item{\code{"data.name"}}{a character string giving the names of the data.}
  \item{\code{"call"}}{the matched call.}
}}
}
\description{
Test for association between paired samples, using one of Pearson's product moment correlation coefficient,Kendall's \eqn{\tau}{tau} or Spearman's \eqn{\rho}{rho}.
Unlike the stats version of cor.test, this function allows users to set the null to a value other than zero.
}
\references{
TBA
}
