% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_smd_calc.R
\name{boot_smd_calc}
\alias{boot_smd_calc}
\alias{boot_smd_calc.default}
\alias{boot_smd_calc.formula}
\title{Bootstrapped SMD Calculation}
\usage{
boot_smd_calc(
  x,
  ...,
  paired = FALSE,
  var.equal = FALSE,
  alpha = 0.05,
  bias_correction = TRUE,
  rm_correction = FALSE,
  glass = NULL,
  boot_ci = c("stud", "basic", "perc"),
  R = 1999
)

\method{boot_smd_calc}{default}(
  x,
  y = NULL,
  paired = FALSE,
  var.equal = FALSE,
  alpha = 0.05,
  mu = 0,
  bias_correction = TRUE,
  rm_correction = FALSE,
  glass = NULL,
  boot_ci = c("stud", "basic", "perc"),
  R = 1999,
  ...
)

\method{boot_smd_calc}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{alpha}{alpha level (default = 0.05)}

\item{bias_correction}{Apply Hedges' correction for bias (default is TRUE).}

\item{boot_ci}{type of bootstrap confidence interval. Options include studentized (stud), empirical/basic (basic) and percentile (perc) confidence intervals.}

\item{R}{number of bootstrap replicates}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{mu}{Null value. Deviating from zero will give the x-y-mu.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}
}
\value{
A data frame containing the SMD estimates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

A function to only calculate standardized mean differences with bootstrap confidence intervals.
}
\details{
For details on the calculations in this function see vignette("SMD_calcs").
}
\examples{
\dontrun{
boot_smd_calc(formula = extra ~ group,data = sleep, paired = TRUE, smd_ci = "nct")
}
}
\seealso{
Other effect sizes: 
\code{\link{ses_calc}()},
\code{\link{smd_calc}()}
}
\concept{effect sizes}
