% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_t_test.R
\name{boot_t_test}
\alias{boot_t_test}
\alias{boot_t_test.default}
\alias{boot_t_test.formula}
\title{Bootstrapped t-test}
\usage{
boot_t_test(x, ...)

\method{boot_t_test}{default}(
  x,
  y = NULL,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater", "equivalence", "minimal.effect"),
  mu = 0,
  alpha = 0.05,
  R = 1999,
  ...
)

\method{boot_t_test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", "less", "equivalence" (TOST), or "minimal.effect" (TOST). You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean for the two tailed test (or difference in means if you are performing a two sample test).}

\item{alpha}{alpha level (default = 0.05)}

\item{R}{number of bootstrap replicates}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
  \item{\code{statistic}}{the value of the t-statistic.}
  \item{\code{parameter}}{the degrees of freedom for the t-statistic.}
  \item{\code{p.value}}{the p-value for the test.}
  \item{\code{conf.int}}{a confidence interval for the mean appropriate to the specified alternative hypothesis.}
  \item{\code{estimate}}{the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.}
  \item{\code{null.value}}{the specified hypothesized value of the mean or mean difference. May be 2 values.}
  \item{\code{stderr}}{the standard error of the mean (difference), used as denominator in the t-statistic formula.}
  \item{\code{alternative}}{a character string describing the alternative hypothesis.}
  \item{\code{method}}{a character string indicating what type of t-test was performed.}
  \item{\code{data.name}}{a character string giving the name(s) of the data..}
}
}
\description{
A function for a bootstrap method for Tt-tests.
}
\details{
For details on the calculations in this function see vignette("robustTOST").

The implemented test(s) corresponds to the proposal of Chapter 16 of Efron and Tibshirani (1994).

For details on the calculations in this function see vignette("robustTOST").
}
\section{References}{


Efron, B., & Tibshirani, R. J. (1994). An introduction to the bootstrap. CRC press.
}

\seealso{
Other Robust TOST: 
\code{\link{boot_log_TOST}()},
\code{\link{boot_t_TOST}()},
\code{\link{log_TOST}()},
\code{\link{wilcox_TOST}()}
}
\concept{Robust TOST}
