% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunner_munzel.R
\name{brunner_munzel}
\alias{brunner_munzel}
\alias{brunner_munzel.default}
\alias{brunner_munzel.formula}
\title{Brunner-Munzel Test}
\usage{
brunner_munzel(
  x,
  alternative = c("two.sided", "less", "greater"),
  p_null = 0.5,
  alpha = 0.05,
  ...
)

\method{brunner_munzel}{default}(
  x,
  y = NULL,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  p_null = 0,
  alpha = 0.05,
  ...
)

\method{brunner_munzel}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{p_null}{a number specifying an optional parameter used to form the null hypothesis (Default = 0.5). This can be thought of as the null in terms of probability of superiority, p = P (X < Y ) + 0.5 * P (X = Y); See ‘Details’.}

\item{alpha}{alpha level (default = 0.05)}

\item{...}{further arguments to be passed to or from methods.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").}
}
\value{
A list with class \code{"htest"} containing the following components:
\describe{
  \item{\code{statistic}}{the value of the test statistic.}
  \item{\code{parameter}}{the degrees of freedom for the test statistic.}
  \item{\code{p.value}}{the p-value for the test.}
  \item{\code{conf.int}}{a confidence interval for the mean appropriate to the specified alternative hypothesis.}
  \item{\code{estimate}}{the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.}
  \item{\code{null.value}}{the specified hypothesized value of the mean or mean difference. May be 2 values.}
  \item{\code{stderr}}{the standard error.}
  \item{\code{alternative}}{a character string describing the alternative hypothesis.}
  \item{\code{method}}{a character string indicating what type of test was performed.}
  \item{\code{data.name}}{a character string giving the name(s) of the data.}
}
}
\description{
This is a generic function that performs a generalized asymptotic Brunner-Munzel test in a fashion similar to \code{"t.test"}
}
\details{
The estimate of probability of superiority refers to the following:

 p(X<Y) + .5*P(X=Y)
}
\examples{
data(mtcars)
#brunner_munzel(mpg ~ am,
#data = mtcars)
}
\references{
reference
Brunner, E., Munzel, U. (2000). The Nonparametric Behrens-Fisher Problem: Asymptotic Theory and a Small Sample Approximation. Biometrical Journal 42, 17 -25.

Neubert, K., Brunner, E., (2006). A Studentized Permutation Test for the Nonparametric Behrens-Fisher Problem. Computational Statistics and Data Analysis.

Munzel, U., Brunner, E. (2002). An Exact Paired Rank Test. Biometrical Journal 44, 584-593.
}
\seealso{
Other TOST: 
\code{\link{simple_htest}()},
\code{\link{t_TOST}()},
\code{\link{tsum_TOST}()}
}
\concept{TOST}
