% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_smd.R
\name{compare_smd}
\alias{compare_smd}
\title{Comparing SMDs between independent studies}
\usage{
compare_smd(
  smd1,
  n1,
  smd2,
  n2,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  null = 0,
  TOST = FALSE
)
}
\arguments{
\item{smd1}{SMD from study 1.}

\item{n1}{sample size(s) from study 1 (can be 1 number or vector of 2 numbers).}

\item{smd2}{SMD from study 2.}

\item{n2}{sample size(s) from study 2 (can be 1 number or vector of 2 numbers).}

\item{paired}{a logical indicating whether the SMD is from a paired or independent samples design. If a one-sample design, then paired must be set to TRUE.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{null}{a number indicating the null hypothesis. For TOST, this would be equivalence bound.}

\item{TOST}{logical indicator (default = FALSE) to perform two one-sided tests of equivalence (TOST). Minimal effects testing not currently available. If specified, alternative is ignored.}
}
\value{
A list with class "htest" containing the following components:
\describe{
  \item{\code{"statistic"}}{z-score}
  \item{\code{"p.value"}}{numeric scalar containing the p-value for the test under the null hypothesis.}
  \item{\code{"estimate"}}{difference in SMD between studies}
  \item{\code{"null.value"}}{the specified hypothesized value for the null hypothesis.}
  \item{\code{"alternative"}}{character string indicating the alternative hypothesis (the value of the input argument alternative). Possible values are "greater", "less", or "two-sided".}
  \item{\code{"method"}}{Type of SMD}
  \item{\code{"data.name"}}{"Summary Statistics" to denote summary statistics were utilized to obtain results.}
  \item{\code{"smd"}}{SMDs input for the function.}
  \item{\code{"sample_sizes"}}{Sample sizes input for the function.}
  \item{\code{"call"}}{the matched call.}
}
}
\description{
A function to compare standardized mean differences (SMDs) between studies. This function is intended to be used to compare the compatibility of original studies with replication studies (lower p-values indicating lower compatibility)
}
