% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatostpaired.h.R
\name{dataTOSTpaired}
\alias{dataTOSTpaired}
\title{TOST Paired Samples T-Test}
\usage{
dataTOSTpaired(data, pairs, lowEqBDz = -0.5, highEqBDz = 0.5,
  alpha = 0.05, desc = FALSE, plots = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{pairs}{a list of vectors of strings naming variables to pair from 
\code{data}}

\item{lowEqBDz}{a number (default: -0.5), the lower equivalence bound in 
Cohen's Dz}

\item{highEqBDz}{a number (default: 0.5), the upper equivalence bound in 
Cohen's Dz}

\item{alpha}{a number between 0 and 1 (default: 0.05) specifying the alpha 
level}

\item{desc}{\code{TRUE} or \code{FALSE} (default), provide descriptive 
statistics}

\item{plots}{\code{TRUE} or \code{FALSE} (default), provide plots}
}
\description{
TOST Paired Samples T-Test
}
