% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd_plot.R
\name{plot_smd}
\alias{plot_smd}
\title{Function to produce plots of the distribution of standardized mean difference}
\usage{
plot_smd(
  lambda,
  df,
  SE,
  smd_label,
  ci_shade = c(0.5, 0.9, 0.95, 0.999),
  ci_lines = c(0.9, 0.95)
)
}
\arguments{
\item{lambda}{The non-centrality parameter for the standardized mean difference}

\item{df}{Degrees of freedom for the standardized mean difference}

\item{SE}{The standard error of the for the standardized mean difference (square root of the variance of the SMD)}

\item{smd_label}{The label that is desired for the plot}

\item{ci_shade}{At which confidence intervals to change the color shade of the distribution}

\item{ci_lines}{The confidence intervals to display on the line at the bottom of the plots}
}
\value{
Returns plot of the distribution of the standardized mean difference.
}
\description{
Function to produce plots of the distribution of standardized mean difference
}
\details{
This function was created so that users could create plots from their own SMD calculations.
The difficulty is that specific information must be past onto this function.
The non-centrality parameter (lambda), degrees of freedom (df), and standard error (SE) calculations can be found in the vignettes of this package.
}
