% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_z_cor.R
\name{power_z_cor}
\alias{power_z_cor}
\title{Power Calculations for Correlations}
\usage{
power_z_cor(
  n = NULL,
  rho = NULL,
  power = NULL,
  null = 0,
  alpha = NULL,
  alternative = c("two.sided", "less", "greater", "equivalence")
)
}
\arguments{
\item{n}{number of observations.}

\item{rho}{true correlation value (alternative hypothsis).}

\item{power}{statistical power (1-beta).}

\item{null}{the null hypothesis value.}

\item{alpha}{a priori alpha-level (i.e., significance level).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater", "less", or "equivalence" (TOST). You can specify just the initial letter.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Calculates the approximate power for a z-test based on a correlation.
}
\references{
Lakens, D. (2017). Equivalence Tests: A Practical Primer for t Tests, Correlations, and Meta-Analyses. Social Psychological and Personality Science, 8(4), 355–362. https://doi.org/10.1177/1948550617697177

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd Ed). Hillsdale, NJ: Lawrence Erlbaum Associates.

Zhang, Z., & Yuan, K.-H. (2018). Practical Statistical Power Analysis Using Webpower and R (Eds). Granger, IN: ISDSA Press.
}
\seealso{
Other power: 
\code{\link{power_eq_f}()},
\code{\link{power_t_TOST}()}
}
\concept{power}
