% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOSTone.raw.R
\name{powerTOSTone.raw}
\alias{powerTOSTone.raw}
\title{Power analysis for TOST for one-sample t-test (raw scores)}
\usage{
powerTOSTone.raw(alpha, statistical_power, sd, low_eqbound, high_eqbound,
  epsilon)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in raw scores}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in raw scores}

\item{epsilon}{By default, true effect is assumed to be 0. If you want to perform an equivalence test when expecting a non-zero effect, specify the expected true effect raw scores as epsilon (e.g., 0.05)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of observations needed in each group
}
\description{
Power analysis for TOST for one-sample t-test (raw scores)
}
\examples{
powerTOSTone.raw(alpha=0.05, statistical_power=0.9, low_eqbound=-0.3, high_eqbound=0.3, epsilon=0)
}

