% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerTOSTpaired.raw.R
\name{powerTOSTpaired.raw}
\alias{powerTOSTpaired.raw}
\title{Power analysis for TOST for dependent t-test (raw scores)}
\usage{
powerTOSTpaired.raw(alpha, statistical_power, low_eqbound, high_eqbound, sdif,
  epsilon)
}
\arguments{
\item{alpha}{alpha used for the test (e.g., 0.05)}

\item{statistical_power}{desired power (e.g., 0.8)}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's dz)}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's dz)}

\item{sdif}{standard deviation of the difference scores}

\item{epsilon}{By default, true effect is assumed to be 0. If you want to perform an equivalence test when expecting a non-zero effect, specify the expected standardized effect size in Cohen's dz as epsilon (e.g., 0.05)}
}
\value{
Returns a string summarizing the power analysis, and a numeric variable for the number of pairs needed
}
\description{
Power analysis for TOST for dependent t-test (raw scores)
}
\examples{
powerTOSTpaired.raw(alpha=0.05,statistical_power=0.8,low_eqbound=-3, high_eqbound=3, sdif=10, epsilon=0)
}

