% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.TOSTnp.R
\name{TOSTnp-methods}
\alias{TOSTnp-methods}
\alias{print.TOSTnp}
\title{Methods for TOSTnp objects}
\usage{
\method{print}{TOSTnp}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class \code{TOSTnp} as returned from the reli_stats function}

\item{digits}{Number of digits to print for p-values}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{TOSTnp-methods}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the Limits of Agreement}
}
}
\description{
Methods defined for objects returned from the agree functions.
}
