% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTone.raw.R
\name{TOSTone.raw}
\alias{TOSTone.raw}
\title{TOST function for a one-sample t-test (raw scores)}
\usage{
TOSTone.raw(m, mu, sd, n, low_eqbound, high_eqbound, alpha)
}
\arguments{
\item{m}{mean}

\item{mu}{value to compare against}

\item{sd}{standard deviation}

\item{n}{sample size}

\item{alpha}{alpha level (default = 0.05)}

\item{low_eqbound_d}{lower equivalence bounds (e.g., -0.5) expressed in standardized mean difference (Cohen's d)}

\item{high_eqbound_d}{upper equivalence bounds (e.g., 0.5) expressed in standardized mean difference (Cohen's d)}
}
\value{
Returns TOST t-value 1, TOST p-value 1, TOST t-value 2, TOST p-value 2, degrees of freedom, low equivalence bound, high equivalence bound, Lower limit confidence interval TOST, Upper limit confidence interval TOST
}
\description{
TOST function for a one-sample t-test (raw scores)
}
\examples{
TOSTone.raw(m=0.52,mu=0.5,sd=0.5,n=300,low_eqbound=-0.1, high_eqbound=0.1, alpha=0.05)
}

