% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOSTpaired.raw.R
\name{TOSTpaired.raw}
\alias{TOSTpaired.raw}
\title{TOST function for a dependent t-test (raw scores)}
\usage{
TOSTpaired.raw(n, m1, m2, sd1, sd2, r12, low_eqbound, high_eqbound, alpha)
}
\arguments{
\item{n}{sample size (pairs)}

\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{r12}{correlation of dependent variable between group 1 and  group 2}

\item{low_eqbound}{lower equivalence bounds (e.g., -0.5) expressed in raw scores}

\item{high_eqbound}{upper equivalence bounds (e.g., 0.5) expressed in raw scores}

\item{alpha}{alpha level (default = 0.05)}
}
\value{
Returns dataframe with t-value, degrees of freedom, p-value, and lower and upper limit of confidence interval
}
\description{
TOST function for a dependent t-test (raw scores)
}
\examples{
TOSTpaired.raw(n=65,m1=5.830769,m2=5.7538462,sd1=1.1668498,sd2=1.2994081,r12=0.744992,low_eqbound=-0.337341706,high_eqbound=0.337341706, alpha=0.05)
}

