% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.TOSTt.R
\name{TOSTt-methods}
\alias{TOSTt-methods}
\alias{print.TOSTt}
\alias{plot.TOSTt}
\title{Methods for TOSTt objects}
\usage{
\method{print}{TOSTt}(x, digits = getOption("digits"), ...)

\method{plot}{TOSTt}(x, type = "c", ci_lines, ci_shades, ...)
}
\arguments{
\item{x}{object of class \code{TOSTt} as returned from the reli_stats function}

\item{digits}{Number of digits to print for p-values}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{TOSTt-methods}}.}

\item{type}{Type of plot to produce. Default is a consonance plot "c" but consonance distribution plot can be produced with "cd".}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the Limits of Agreement}
  \item{\code{plot}}{Returns a plot of the data points used in the reliability analysis}
}
}
\description{
Methods defined for objects returned from the agree functions.
}
